fixed the bug with the state then there is no internet on DetailsScreen

+++ app/src/main/java/com/example/pexelsapp/screens/details/DetailsScreenViewModel.kt
@@ -19,12 +19,10 @@ class DetailsScreenViewModel @Inject constructor(
     private val _uiState = MutableStateFlow<DetailsScreenUiState>(DetailsScreenUiState.Loading)
     val uiState = _uiState.asStateFlow()
 
-    private val _state = MutableStateFlow<DetailsScreenState?>(null)
-    val state = _state.asStateFlow()
-
     fun onDownloadClick() {
-        val prevState = _state.value ?: return
-        downloader.downloadImage(prevState.photo.url)
+        val prevState = _uiState.value
+        if (prevState is DetailsScreenUiState.Loaded)
+            downloader.downloadImage(prevState.photo.url)
     }
 
     fun getPhotoById(photoId: Long) {
@@ -34,11 +32,14 @@ class DetailsScreenViewModel @Inject constructor(
                 val photo = repository.getPhotoById(photoId)
                 val savedPhoto = repository.getFavouritePhotoById(photoId)
                 if (savedPhoto == null) {
-                    _state.value = DetailsScreenState(photo!!, false)
+                    if (photo != null) {
+                        _uiState.value = DetailsScreenUiState.Loaded(photo, false)
+                    } else {
+                        _uiState.value = DetailsScreenUiState.NotFound
+                    }
                 } else {
-                    _state.value = DetailsScreenState(savedPhoto, true)
+                    _uiState.value = DetailsScreenUiState.Loaded(savedPhoto, true)
                 }
-                _uiState.value = DetailsScreenUiState.Loaded
             } catch (e: Exception) {
                 _uiState.value = DetailsScreenUiState.NotFound
             }
@@ -47,14 +48,16 @@ class DetailsScreenViewModel @Inject constructor(
 
     fun onBookmarkSavedStateChanged() {
         viewModelScope.launch {
-            val currentState = _state.value ?: return@launch
-            val savedPhoto = repository.getFavouritePhotoById(currentState.photo.id)
-            if (savedPhoto == null) {
-                repository.saveFavouritePhoto(currentState.photo)
-                _state.value = currentState.copy(isFavourite = true)
-            } else {
-                repository.deleteFavouritePhoto(currentState.photo)
-                _state.value = currentState.copy(isFavourite = false)
+            val currentState = _uiState.value
+            if (currentState is DetailsScreenUiState.Loaded) {
+                val savedPhoto = repository.getFavouritePhotoById(currentState.photo.id)
+                if (savedPhoto == null) {
+                    repository.saveFavouritePhoto(currentState.photo)
+                    _uiState.value = currentState.copy(isFavourite = true)
+                } else {
+                    repository.deleteFavouritePhoto(currentState.photo)
+                    _uiState.value = currentState.copy(isFavourite = false)
+                }
             }
         }
     }
